+++ FeetOfAzeroth/FeetOfAzerothText.lua
@@ -23,18 +23,22 @@ fa.MAIN_SESSION = "Session : "
 fa.MAIN_GPS = "GPS : "
 fa.MAIN_SPD = "Vitesse "
 
--- fenêtre about
+-- fenêtre à propos
 
 fa.ABT_VERSION = "Version"
 fa.ABT_BY = "By "
 
 -- fenêtre des stats
 
-fa.STAT_DIST_TOT = "Distance totale : "
-fa.STAT_DIST_SESS = "Distance session : "
+fa.STAT_DIST_TOT = "Distance totale : %.2f yds"
+fa.STAT_DIST_SESS = "Distance session : %.2f yds"
 fa.STAT_PH = "Noodle soup & naps keep you healthy !"
+fa.STAT_AFK = "AFK : "
+fa.STAT_GHOST = "Ghost : "
+fa.STAT_VITYD = "Record de vitesse : %.2f yd/s"
+fa.STAT_VITKM = "Record de vitesse : %.2f km/h"
 
--- messages console
+-- messages dans la console
 
 fa.CSL_READY = "Ready to walk !"
 fa.CSL_FEATS_OK = "Succès validé !"